+++ cloud/cloud_test.go
@@ -31,7 +31,7 @@ func TestResourceTypePluralizeName(t *testing.T) {
 		{in: "group", out: "groups"},
 		{in: "policy", out: "policies"},
 		{in: "internetgateway", out: "internetgateways"},
-		{in: "registry", out: "registries"},
+		{in: "repository", out: "repositories"},
 	}
 	for _, tc := range tcases {
 		if got, want := PluralizeResource(tc.in), tc.out; got != want {