+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 ## Features
 
 - `awless ls policies` now returns: your managed policies + all policies attached to any users, role or group
-- Listing [ECR](https://aws.amazon.com/ecr/) repositories: `awless list registries`
-- Create/Delete [ECR](https://aws.amazon.com/ecr/) repositories: `awless create/delete registry`
+- Listing [ECR](https://aws.amazon.com/ecr/) repositories: `awless list repositories`
+- Create/Delete [ECR](https://aws.amazon.com/ecr/) repositories: `awless create/delete repository`
 
 ## v0.1.0 [2017-05-31]